complete extraction, transformation and loading of data with SDK
