--- transform-lambda/test/test.csv
@@ -1,3 +1,0 @@
-a,b,c,d,e,f,g
-1,2,3,4,5,6,7
-8,9,10,11,12,13
